refactor: move utils dir
