+++ main/uis/Styled/StyledComponents.tsx
@@ -1,6 +1,6 @@
 import {Animated} from 'react-native';
 import type {DoobooTheme} from '@dooboo-ui/theme';
-import {isEmptyObject} from '../../utils';
+import {isEmptyObject} from '../../utils/utils';
 import {light} from '@dooboo-ui/theme';
 import styled from '@emotion/native';
