+++ main/uis/Typography/TypographyInverted.tsx
@@ -1,7 +1,7 @@
 import {light, withTheme} from '@dooboo-ui/theme';
 
 import type {DoobooTheme} from '@dooboo-ui/theme';
-import {isEmptyObject} from '../../utils';
+import {isEmptyObject} from '../../utils/utils';
 import styled from '@emotion/native';
 
 // Title
